Add back simulated input
This commit re-adds the ability to use simulated input.  In doing so it
re-structures the input classes.  Now there is an interface for use by
all forms of input (IGameInput); this is now the type of
GameSetup.Input.

The BellowsInput abstract class still exists.  Currently its only
subclass is TcpInput, but in the future other subclasses could be added
(e.g. a class for a direct serial connection to the bellows).  TcpInput
is used for both actual bellows data and data from a text file.  This
is because developers need some way to test these classes without
needing to be connected to an MRI.  Moreover, the data in the text
files is based on actual data, so it should act as close to an actual
connection to an MRI as possible.

The simulated input is handled by a new class SimInput that implements
the IGameInput interface.

Which input class (SimInput or TcpInput) is used is controlled by the
presence of a config file.  If the config file exists and has a valid
IP address, TcpInput is use.  Else SimInput is used.  A config file was
re-established in order to get rid of hard coded IP addresses, and
allow users to switch between TcpInput and SimInput without needing to
modify code.  However, the location of the new file was moved so as to
allow a build executable to find/use it.