Backup Server Changes
Note: This commit is not stable.  It was done to backup the changes I
have been working on.  However, this branch was not stable to begin
with, and the simulated input functionality still works, so this commit
does not break anything that was previously working.

Once stable, the changes will be merged into master.