Two Updates
Log: Changed bellowslog file to gamedatalog to reflect that game data,
not bellows data is being logged to it.

GameServer: All the sample files we were given contain “game data” in
that it has already been subtracted from 4095.  GameServer was updated
to account for this.  The change may seem kludgey, but was done this
way after serious thought:
- We want to be able to use gamedatalog files from actual scans as test
data, so GameServer needs to use data in this format.
- The game should be logging the data it used (i.e. game data) not
bellows data.  If we wanted to log bellows data, the GameServer should
do this, not the Input module.
- The TcpInput module should not have to distinguish between data from
a file or actual bellows data.
- The subtraction from 4095 for actual bellows data needs to happen in
TcpInput, as there is the possibility of a direct serial connection
being added.  This new class will need to do the subtraction (since
there is no GameServer to do it for it), and the TcpInput and
SerialInput classes should work the same.