Adjustments to player and pick-up positioning
Prior to playing the game, there is a calibration which maps the max
and min bellows positions to the character’s range of motion.  However,
during the game the player can generate bellows positions outside this
range:
- They cough
- The bellows starts to slide/changes position
- There is a change in their breathing

To deal with this, two changes where made:
- PlayerController was updated to smooth out and better handle data
outside the range of the calibration’s max and min.
- PowerUpSpawner was changed to shift where the pick-ups spawn so they
are closer to the center.  The idea is that even if “calibration drift”
no longer allows the patient to get the avatar all the way to the
bottom of the screen, they will still be able to collect the pickups
when they have fully exhaled.